Refactor ValidationUtil out of Plugin directory
