+++ arelle/plugin/validate/ESEF/ESEF_Current/ValidateXbrlFinally.py
@@ -26,7 +26,7 @@ from arelle.ModelValue import QName
 from arelle.ModelValue import qname
 from arelle.ModelXbrl import ModelXbrl
 
-from arelle.plugin.validate.Util import etreeIterWithDepth
+from arelle.utils.validate.ValidationUtil import etreeIterWithDepth
 from arelle.PythonUtil import isLegacyAbs, normalizeSpace
 from arelle.PythonUtil import strTruncate
 from arelle.utils.validate.DetectScriptsInXhtml import containsScriptMarkers
