+++ arelle/utils/validate/ValidationUtil.py
